requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
tokenizers>=0.15.0

# Configuration
omegaconf>=2.3.0
hydra-core>=1.3.0
pyyaml>=6.0

# ML
scikit-learn>=1.3.0
numpy>=1.24.0

# Quantization (optional)
bitsandbytes>=0.41.0

# Data processing
datasets>=2.14.0
pandas>=2.0.0

# API clients
dashscope>=1.14.0  # For Qwen API
openai>=1.0.0      # For OpenAI-compatible APIs

# Pipeline (optional)
dvc>=3.0.0

# Development
tqdm>=4.65.0
